[feat] Manager 생성으로 구조 변경
