+++ test-shell/src/main/java/IManager.java
@@ -1,0 +1,13 @@
+public interface IManager {
+    String read(int address);
+
+    boolean write(int address, String value);
+
+    boolean fullread();
+
+    boolean fullwrite(String value);
+
+    boolean erase(int address, int size);
+
+    boolean erase_range(int startLBA, int endLBA);
+}
